pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3", "wheel", "pybind11>=2.6.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sam-studio"
version = "1.0.0"
authors = [
  {name = "Amir Kimiyaie", email = "Amirrad3010@gmail.com"}
]
description = "SAM Labeling Studio: AI-assisted annotation tool"
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
"PyQt6==6.5.0",
"Pillow>=9.4.0",
"numpy>=1.24.4",
"pyyaml",
"requests==2.28.0",
"pydantic",
"fastapi",
"python-multipart",
"aiohttp",
"smallest_kgon @ git+https://github.com/entechlab/smallest-k-gon-python.git",

# SAM
"torch>=2.5.1",
"torchvision>=0.20.1",
"tqdm>=4.66.1",
"hydra-core>=1.3.2",
"iopath>=0.1.10",
"uvicorn"
]

[project.urls]
Homepage = "https://github.com/entechlab/sam-labeling-studio"

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]

[tool.ruff]
line-length = 102

[tool.ruff.lint]
# line too long
extend-select = ["E501"]

